ubi: Add host_fw_watch
Create a new watch function to monitor for a host-fw patch file
that is uploaded for the user. Once detected, this file, which
would be a squashfs, will be loaded into a ubi volume. The reason
for being a squashfs is that the lid files exceed the size of the
NOR chip so it's required to have them in a compressed form.

Change-Id: I189bd6fd2db94a83c98170c434013e21b133edbf
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>