ubi: Add callback function for host_fw_watch
When a file is detected by the host_fw_watch, call the ubipatch
service. This service will be enhanced to mount the host-fw path
volume when a patch file is uploaded. Calling this service file
doesn't affect the pnor patch because it's already mounted. The
advantage of using an existing service file is that there's no
dependencies with adding a new service file to the recipe, since
patching of the host will change once eMMC systems are available.

Change-Id: I00806e5cabe18a437ebb16a838393608d1d68b58
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>