ubi: Mount host-fw patch volume
The ubipatch service is called when the BMC boots and when a
host-fw patch file is detected.
1. Add a mkdir command to ubipatch so that the directory where the
patch file is to be uploaded exists. This would be a tmpfs directory.
2. If a host-fw patch file exists, remove the previous patch volume
if it exists, then load the new patch file into a read-only ubi
volume, then delete the file to free up space.
3. Remove the host-fw patch from the ubimount service, so that if a
pnor-only update is requested, space is freed up. This means that
the user wants to run from the pnor version and not from the code
in host-fw.

Tested:
- Uploading host-fw files in a row deletes the previous one and mounts
the new one.
- A pnor update removes the host-fw volume.
- After a bmc reboot, the host-fw volume is remounted.

Change-Id: I675516a41558aec3f25caf0f2ea34f87132349ac
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>