configure: Make ubi an opt-in option instead of opt-out
Adding a 3rd option (mmc) makes the combinations complex to
setup via a recipe file the way that it's currently setup
where you have to opt-out from the default ubi.

Change it so that the default is static, then ubi and mmc
are opt-in options. A recipe update will be done to bump
to this change.

Ustream we'll create an option to be able to choose just
one of the configurations excluding the rest.

Change-Id: Ic73024140e4d433bbe562027e4290d3059e3a67f
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>