Add a mount host-fw service for eMMC
An external build tool will append an ext4 filesystem with
the most up-to-date host firmware files to the wic eMMC image
file. This method is simpler than trying to modify the wic
image file to insert files.

If this partition exists, mount it to /usr/share/host-fw/
(where the original files reside) so that they take effect.
This is faster than copying the files, since they don't need
to be modified anyways.

This service will not be needed on the final implementation
of the eMMC layout.

Change-Id: Ib12a3900d983cef3ee38a213f76369703279cd05
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>