vpnor: Dynamically find hostfw partition
The size of the wic file slightly grew, causing the offset to the
hostfw partition to change. To adapt to future changes, attempt to
dynamically find the hostfw partition.

This is temporary as the final eMMC layout will have the hostfw
code out of the BMC filesystem and into a separate device.

Tested: Verified the hostfw partition is mounted with the latest
        OpenBMC build.

Fixes: SW494717

Change-Id: I88ab664d532a088f09609093cb7fe551cf98d969
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>