Expose HBB lid over PCIe space (#5)
Created a service to load the Host Boot Bootloader lid file on to the
memory mapped region of PCIe address space.Hostboot need the HBB lid
file to be loaded during the initial boot. Currently mbox is used to
transfer the HBB lid file to hostboot. Going forward we are planning to
deprecate Mbox and use Memory mapped PCIe address space for the HBB lid.
This address is provided to to the host to load the file.

Tested: Verified the sha256 hash of raw memory region of exposed lid
file at required address via /dev/mem with the original lid file.

Change-Id: Id25008bbf67738b491f59fe28f87c9dbb6112f7c

Signed-off-by: Manjeet Singh <itsmanjeet1998@gmail.com>