add function mmc_get_primary_label to update logic for bootside check
Currently bootside is used to mount rofs but if BMC fails to boot the
bootside env variable is not updated properly. This fix changes the
logic of checking the bootside to properly mount rofs regardless of
bootside value.

Tested: Tested by corrupting kernel then performing a system reboot.
Checking the fw_printenv output for bootside and comparing with the
outputs of df and mount shows that this fix mounts the correct bootside
regardless of the bootside env variable value.

Change-Id: I99bcf5e5abc4483b010cccb1c47082948e096ec1
Signed-off-by: Bryon Catlin <bryon.catlin@gmail.com>